[Refactor/#390] 이메일 전송 배치 전송 설정에 따라 조회하도록 수정 (#391)
* feat: DayCode 추가

* refactor: 이메일 전송 배치 전송 설정에 따라 조회하도록 수정
